tests/submission/W3C/domhtml/HTMLLinkElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>W3C</title>
<link href="/support/html/sample.html" hreflang="en" media="screen" rel="Glossary" type="text/html">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-link-element.html#dom-link-href" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<p>The href attribute specifies the URI of the linked resource.     

    Retrieve the href attribute and examine its value.</p>


<script>
var domtest = async_test("The href attribute specifies the URI of the linked resource.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vhref;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("link");
      assert_equals(nodeList.length,3);
      testNode = nodeList.item(0);

      assert_uri_equals(document.location.protocol,null,document.location.host,"sample.html",null,null,null,true,testNode.href);

    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
